Add separation of express App & server.
