+++ server/config/test.yml
@@ -1,0 +1,32 @@
+database:
+  secret: 'keyboard cat'
+  config:
+    # This is needed bc of the docker-compose file
+    authSource: 'admin'
+    auth:
+      user: 'root'
+      password: 'example'
+    authMechanism: 'SCRAM-SHA-1'
+
+mailing:
+  testingMode: false
+  templates: 
+  # Replacements:
+  # {{name}}: Name of user
+  # {{username}}: Username of user
+  # {{password}}: Temporary password of user
+    credentials: "Hallo {{name}},\n
+      \n
+      hier sind deine Zugangsdaten zum Tutor-Management-System:\n
+      \n
+      Nutzername: {{username}}\n
+      Passwort: {{password}}\n
+      \n
+      Mit freundlichen Grüßen\n
+      TMS Admin"
+  host: 'SERVER'
+  port: 587
+  from: 'MAIL'
+  auth:
+    user: USER
+    pass: PASS
